refactor(decompression): move animated vec3 unpacking to its own header
